Converte componentes de JSX para TSX e adiciona tipos
