--- src/components/Message/Message.jsx
@@ -1,14 +1,0 @@
-import styles from "./Message.module.css"
-import * as PropTypes from "prop-types";
-
-export function Message({text}){
-    return (
-        <div className={styles.messageContainer}>
-            <h1>{text}</h1>
-        </div>
-    );
-}
-
-Message.propTypes = {
-    texto: PropTypes.string.isRequired
-}
