--- src/data/PartnersData.js
@@ -1,119 +1,0 @@
-import af from '../assets/images/parceiros/2024/af.webp'
-import artway from '../assets/images/parceiros/2024/artway.webp'
-import cdc from '../assets/images/parceiros/2024/cdc.webp'
-import eletrolaser from '../assets/images/parceiros/2024/eletrolaser.webp'
-import feedback from '../assets/images/parceiros/2024/feedback.webp'
-import guedes from '../assets/images/parceiros/2024/guedes.webp'
-import inovatec from '../assets/images/parceiros/2024/inovatec.webp'
-import knex from '../assets/images/parceiros/2024/logo_knex.webp'
-import moinhopatoense from '../assets/images/parceiros/2024/moinhopatoense.webp'
-import nutes from '../assets/images/parceiros/2024/nutes.webp'
-import nutrimassa from '../assets/images/parceiros/2024/nutrimassa.webp'
-import onitelecom from '../assets/images/parceiros/2024/onitelecom.webp'
-import papelariapatoense from '../assets/images/parceiros/2024/papelariapatoense.webp'
-import sertaomaker from '../assets/images/parceiros/2024/sertaomaker.webp'
-import sesi from '../assets/images/parceiros/2024/sesi.webp'
-import uepb from '../assets/images/parceiros/2024/uepb.webp'
-import urbano from '../assets/images/parceiros/2024/urbano.webp'
-import vialeste from '../assets/images/parceiros/2024/via_leste.webp'
-import virtus from '../assets/images/parceiros/2024/virtus.webp'
-
-const partnersData = [
-    {
-        href: "https://www.instagram.com/eletrolaser/?igshid=NTc4MTIwNjQ2YQ%3D%3D",
-        image: eletrolaser,
-        alt: "Eletrolaser"
-    },
-    {
-        href: "https://guedesshopping.com.br/",
-        image: guedes,
-        alt: "Guedes Shopping"
-    },
-    {
-        href: "https://www.instagram.com/moinho_patoense/?igshid=NTc4MTIwNjQ2YQ%3D%3D",
-        image: moinhopatoense,
-        alt: "Moinho Patoense"
-    },
-    {
-        href: "https://www.instagram.com/patoensepapelaria/?igshid=NTc4MTIwNjQ2YQ%3D%3D",
-        image: papelariapatoense,
-        alt: "Patoense Papelaria"
-    },
-    {
-        href: "",
-        image: nutrimassa,
-        alt: "Nutrimassa"
-    },
-    {
-        href: "https://www.instagram.com/inovatecuepb/?hl=pt-br",
-        image: inovatec,
-        alt: "Inovatec"
-    },
-    {
-        href: "https://www.instagram.com/artway_pb/",
-        image: artway,
-        alt: "ArtWay"
-    },
-    {
-        href: "https://www.instagram.com/grupovialeste/",
-        image: vialeste,
-        alt: "Grupo Via Leste"
-    },
-    {
-        href: "https://www.instagram.com/escolasesipb/",
-        image: sesi,
-        alt: "SESI"
-    },
-    {
-        href: "https://www.instagram.com/sertaomaker.patos/",
-        image: sertaomaker,
-        alt: "Sertão Maker"
-    },
-    {
-        href: "https://www.instagram.com/onitelecom_/",
-        image: onitelecom,
-        alt: "Oni Telecom"
-    },
-    {
-        href: "https://www.instagram.com/biatic/",
-        image: feedback,
-        alt: "EsusFeedback"
-    },
-    {
-        href: "https://www.instagram.com/urbanonortepatospb/",
-        image: urbano,
-        alt: "Urbano Norte"
-    },
-    {
-        href: "https://www.instagram.com/_arthurfotos/",
-        image: af,
-        alt: "Arthur Fotos"
-    },
-    {
-        href: "https://www.instagram.com/computacao.ccea/",
-        image: cdc,
-        alt: "Ciencias Da Computação"
-    },
-    {
-        href: "https://www.instagram.com/knexjr/",
-        image: knex,
-        alt: "Knex"
-    },
-    {
-        href: "https://www.instagram.com/nutesuepb/",
-        image: nutes,
-        alt: "Nutes"
-    },
-    {
-        href: "https://www.instagram.com/UEPBOficial/",
-        image: uepb,
-        alt: "UEPB"
-    },
-    {
-        href: "https://www.instagram.com/virtuscursos/",
-        image: virtus,
-        alt: "Virtus"
-    }
-];
-
-export default partnersData;
