+++ src/types/MessageProps.ts
@@ -1,0 +1,4 @@
+export interface MessageProps {
+    text: string;
+  }
+
