+++ src/types/OrganizationProps.ts
@@ -1,0 +1,10 @@
+export interface OrganizationMember {
+    image: string;
+    alt: string;
+  }
+  
+  export interface OrganizationProps {
+    title: string;
+    members: OrganizationMember[];
+  }
+
