+++ src/types/Partner.ts
@@ -1,0 +1,6 @@
+export interface Partner {
+    image: string;
+    alt: string;
+    href: string;
+  }
+
